refactor: reduce form props
