+++ src/lib/constants/featured-listing.ts
@@ -1,0 +1,7 @@
+// Constants for creating/editing a featured listing
+
+export const MAX_BEDROOMS = 20;
+export const MAX_BATHROOMS = 20;
+export const MIN_SQUARE_FEET = 100;
+export const MAX_SQUARE_FEET = 50000;
+export const MAX_DESCRIPTION_WORDS = 50;
